Fixed return type of strnlen(), removed sseek, swrite, added debug stmts
Changed the return type of strnlen() to be size_t instead of int.
Removed sseek and swrite from newlib_backend, as well as added in
  the function prototypes for the 19 syscall wrappers.
Added in debugging print statements into the various calls, and redefined
  debug to print nothing.

+++ user/parlib/inc/debug.h
@@ -14,7 +14,7 @@ typedef void * TRUSTED va_list;
 
 #define va_end(ap)  ((void)0)
 
-int	strnlen(const char *s, size_t size);
+size_t	strnlen(const char *s, size_t size);
 void debugfmt(void (*putch)(int, void**), void **putdat, const char *NTS fmt, ...);
 void vdebugfmt(void (*putch)(int, void**), void **putdat, const char *NTS fmt, va_list);
